refactor: Implement abstract `serialize_message` for Singer writers
